requirements.txt
langchain
pypdf2
openai
streamlit
tiktoken
python-dotenv
faiss-cpu
huggingface_hub
InstructorEmbedding
sentence_transformers